Separate receipt scanning and receipt parsing semantically
