+++ backend/split/services/receipt_parser/__init__.py
@@ -1,0 +1,1 @@
+from split.services.receipt_parser.core import extract_relevant_information
